#21 Refactors challenges models and responses
