+++ src/main/java/mobi/appcent/helium/request/challenge/ListOfChallengesReceiptRequest.java
@@ -7,8 +7,7 @@ import mobi.appcent.helium.common.UrlConstant;
 import mobi.appcent.helium.httpClient.HttpMethod;
 import mobi.appcent.helium.model.Pair;
 import mobi.appcent.helium.request.BaseRequest;
-import mobi.appcent.helium.response.account.AccountsResponse;
-import mobi.appcent.helium.response.hotspot.HotspotChallangesResponse;
+import mobi.appcent.helium.response.challenge.ChallengesResponse;
 import okhttp3.Call;
 
 import java.io.IOException;
@@ -49,14 +48,14 @@ public class ListOfChallengesReceiptRequest extends BaseRequest {
         return this;
     }
 
-    public HotspotChallangesResponse execute() throws IOException {
+    public ChallengesResponse execute() throws IOException {
         ArrayList<Pair> queryParams = new ArrayList<>();
         queryParams.add(Pair.create(FieldConstant.CURSOR, cursor));
         queryParams.add(Pair.create(FieldConstant.MIN_TIME, minTime));
         queryParams.add(Pair.create(FieldConstant.MAX_TIME, maxTime));
         queryParams.add(Pair.create(FieldConstant.LIMIT, limit));
         Call call = client.buildCall(UrlConstant.CHALLENGES_PATH, HttpMethod.GET, queryParams, null, null);
-        Type type = TypeToken.get(HotspotChallangesResponse.class).getType();
+        Type type = TypeToken.get(ChallengesResponse.class).getType();
         return execute(call, type);
     }
 }
